Fix favorite updates without "All files access".
